- Setup sandbox project with Debug and Release to test all engine interactions

+++ FretBuzz/FretBuzzFramework/framework/components/gameobject_components/audio_source.cpp
@@ -121,7 +121,7 @@ namespace ns_fretBuzz
 
 		float AudioSource::getLength() const
 		{
-			return m_pISoundSource->getPlayLength();
+			return (float)m_pISoundSource->getPlayLength();
 		}
 
 		bool AudioSource::getIsLooping() const
